Add import template functionality
