+++ .changeset/gold-rabbits-thank.md
@@ -1,0 +1,5 @@
+---
+"go-web-app": patch
+---
+
+Add an option to download excel import template for DREF (Response) which user can fill up and import.
